refactor: organized routes into pages, separating pages and components
