+++ src/Pages/SignUpPage/SignUpStyled.js
@@ -1,4 +1,4 @@
 import styled from "styled-components";
-import { LoginContainer } from "./SignInStyled";
+import { LoginContainer } from "../SignInPage/SignInStyled";
 
 export const CadastroContainer = styled(LoginContainer)``;
